refactor: correct routing
