+++ src/app/app.module.ts
@@ -1,18 +1,14 @@
 import { NgModule, isDevMode } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { StoreModule } from '@ngrx/store';
 import { EffectsModule } from '@ngrx/effects';
 import { StoreDevtoolsModule } from '@ngrx/store-devtools';
-import { StoreRouterConnectingModule } from '@ngrx/router-store';
+// import { StoreRouterConnectingModule } from '@ngrx/router-store';
 import { HttpClientModule } from '@angular/common/http';
 import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app-routing.module';
 import { CoreModule } from './core/core.module';
-import { MainModule } from './main/main.module';
-import { BookingModule } from './booking/booking.module';
-import { CartModule } from './cart/cart.module';
 import { AppState } from './redux/state/app.state';
 import { UserEffects } from './redux/effects/autn-user.effects';
 
@@ -22,9 +18,6 @@ import { UserEffects } from './redux/effects/autn-user.effects';
   ],
   imports: [
     CoreModule,
-    MainModule,
-    BookingModule,
-    CartModule,
     BrowserModule,
     HttpClientModule,
     AppRoutingModule,
@@ -41,7 +34,7 @@ import { UserEffects } from './redux/effects/autn-user.effects';
     }),
     EffectsModule.forRoot([UserEffects]),
     StoreDevtoolsModule.instrument({ maxAge: 25, logOnly: !isDevMode() }),
-    StoreRouterConnectingModule.forRoot(),
+    // StoreRouterConnectingModule.forRoot(),
   ],
   providers: [],
   bootstrap: [AppComponent],
