+++ src/app/booking/booking-routing.module.ts
@@ -1,8 +1,23 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { BookingPageComponent } from './pages/booking-page/booking-page.component';
+import { Path } from '../shared/enums/router.enum';
+import { PassengersComponent } from './components/passengers/passengers.component';
+import { SummaryPageComponent } from './pages/summary-page/summary-page.component';
+import { FlightsSelectionComponent } from './components/flights-selection/flights-selection.component';
 
-const routes: Routes = [{ path: '', component: BookingPageComponent }];
+const routes: Routes = [
+  {
+    path: '',
+    component: BookingPageComponent,
+    children: [
+      { path: Path.Flights, component: FlightsSelectionComponent },
+      { path: Path.Passengers, component: PassengersComponent },
+      { path: Path.Summary, component: SummaryPageComponent },
+      { path: '**', redirectTo: Path.Flights, pathMatch: 'full' },
+    ],
+  },
+];
 
 @NgModule({
   imports: [RouterModule.forChild(routes)],
