+++ src/app/cart/cart.module.ts
@@ -6,21 +6,8 @@ import { CartRoutingModule } from './cart-routing.module';
 import { SharedModule } from '../shared/shared.module';
 import { UserPageComponent } from './pages/user-page/user-page.component';
 import { TableComponent } from './components/table/table.component';
-import { Path } from '../shared/enums/router.enum';
-import { RouterModule, Routes } from '@angular/router';
 import { CartMainPageComponent } from './pages/cart-main-page/cart-main-page.component';
 
-const routes: Routes = [
-  {
-    path: Path.Cart,
-    component: CartMainPageComponent,
-    children: [
-      { path: Path.Empty, component: CartPageComponent },
-      { path: Path.FlightsHistory, component: UserPageComponent },
-      { path: '**', redirectTo: '', pathMatch: 'full' },
-    ],
-  },
-];
 @NgModule({
   declarations: [
     CartComponent,
@@ -33,7 +20,6 @@ const routes: Routes = [
     CommonModule,
     CartRoutingModule,
     SharedModule,
-    RouterModule.forChild(routes),
   ],
   exports: [],
 })
